refactor: store theme logic for mui in init dir
