+++ src/pages/_document.tsx
@@ -1,6 +1,6 @@
 import Document, { Html, Head, Main, NextScript } from 'next/document';
 import createEmotionServer from '@emotion/server/create-instance';
-import theme from '@/styles/initAppTheme';
+import theme from '@/init/mui/loadMuiTheme';
 import { createEmotionCache } from '@/init/mui/emotion';
 
 // eslint-disable-next-line import/no-default-export
